pyproject.toml
[project]
name = "connect-to-qqq100-index-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.7.0",
    "crewai-tools>=1.7.0",
    "fastapi>=0.124.4",
    "langchain-community>=0.4.1",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.7.17",
    "uvicorn>=0.30.0",
    "yfinance>=0.2.66",
]
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
